refactor(admin-ui): split TimeAgo into separate files and add configurable update intervals
Update intervals: <5min = 1s, 5-15min = 10s, 15min-1h = 30s, >1h = no updates.
Intervals are configurable via opts.updateIntervals.
