+++ packages/admin-ui/src/TimeAgo/toEpochMs.ts
@@ -1,0 +1,11 @@
+import type { TDate } from "./types.js";
+
+export function toEpochMs(date: TDate): number {
+    if (date instanceof Date) {
+        return date.getTime();
+    }
+    if (typeof date === "number") {
+        return date;
+    }
+    return new Date(date).getTime();
+}
