+++ packages/admin-ui/src/TimeAgo/toISOString.ts
@@ -1,0 +1,11 @@
+import type { TDate } from "./types.js";
+
+export function toISOString(date: TDate): string {
+    if (date instanceof Date) {
+        return date.toISOString();
+    }
+    if (typeof date === "number") {
+        return new Date(date).toISOString();
+    }
+    return date;
+}
